Project structure edited. Separated huge main into camera.cpp
