--- handy/packages/camera/package.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>camera</name>
-  <version>0.0.1</version>
-  <description>Camera Node for supplying images from attached camera.</description>
-  <maintainer email="denis.bakin5@gmail.com">Denis Bakin</maintainer>
-  <license>MIT</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>cv_bridge</depend>
-  <depend>foxglove_msgs</depend>
-  <depend>OpenCV</depend>
-  <depend>rclcpp</depend>
-  <depend>sensor_msgs</depend>
-  <depend>std_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>visualization_msgs</depend>
-
-
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
