+++ handy/packages/industrial_camera/include/camera.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "CameraApi.h"
 
 #include <rclcpp/rclcpp.hpp>
